rename sdk -> core
